Update Preliminaries.R
"fullCov <<- load(fullCov_Path)" command loads value "fullCov" to variable fullCov rather than the real fullCov data structure from the file fullCov.RData. Note that "*.RData" files have the variable name they were saved with in them so that when they are loaded the variable name is loaded to the environment as well.