Update Functions.R
The terminal gap at chromosomes  2 and 11 met the criterion of MRG=10 so that the preceding region must have been merged to the region after the gap which didn't exist.As a result,   error "outbound index" kept to be thrown when tried to make IRange (line 74). I guessed such an error on line 60. Now I think the problem is solved and we need to check if the IRanges can be created or not.