Update Preliminaries.R
When normalizing the read from different samples with different size libraries the reads for each genes is divided by the library size of that samples and then all values across samples are multiplied to a 'targetsize' so that the targetsize does not make a change. I left it to the default (8+07 reads).