Update Functions.R
I had forgotten to update the RegionMatsPath variable for the next iteration so that by changing the sample the regionMatsPath didn't get updated.
Updating MCC variable is through RegionMat_Path variable. Hence, it needs to be saved to file.path(paste0(DBDir_Path,"RegionMat_Path.RData")) in order to be loaded as the function NewRegion() starts for downstream process.