Update Functions.R
When the sample in next iteration is a new one, it means its RegionMat_[MCC].RData files are not created then list.files() returns empty character leaving RegionMat_Path variable empty. Creating the path to the next RegionMat_[MCC].RData directly make the path absolute and then there is no need for existence of the file.